+++ react-cors-spa/src/App.js
@@ -50,31 +50,6 @@ function App() {
   );
 }
 
-function APIDisplay() {
-  // Access the client
-  const queryClient = useQueryClient()
-
-  // Queries
-  const query = useQuery({ queryKey: ['todos'], queryFn: getTodos })
-
-  return (
-    <div>
-      <ul>{query.data?.map((todo) => <li key={todo.id}>{todo.title}</li>)}</ul>
-
-      <button
-        onClick={() => {
-          mutation.mutate({
-            id: Date.now(),
-            title: 'Display Data',
-          })
-        }}
-      >
-        Add Todo
-      </button>
-    </div>
-  )
-}
-
 
 const APIResult = () => {
   const [data, setData] = useState(null);
