+++ react-cors-spa/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
-import { APIDisplay } from "./App";
+import { APIDisplay } from "./APIDisplay";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
